requirements.txt
# Core
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
huggingface-hub>=0.19.0

# Retrieval & NLI
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Evaluation
rouge-score>=0.1.2
mauve-text>=0.3.0
numpy>=1.24.0
scipy>=1.10.0

# LLM APIs (optional, for commercial models)
openai>=1.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
jsonlines>=3.1.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
